requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
coverage==7.10.7
Django==5.2.6
django-bootstrap5==25.2
django-debug-toolbar==6.0.0
django-extensions==4.1
djangorestframework==3.16.1
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
